Update: package.json, update: index.html, update: index,js
Update package.json installing the socket.io from the terminal.
Update index.html creating the script source from the socket.io/socket.io.js and another script with a variable called socket that create a new instance of the io, if the word connect if finded will appear a console log saying that the socket is connected to the server, and if the word disconnect is finded will appear a console log saying that the socket lost the connection with the server.
Update index.js adding the use properties of socket.io, creating the node server, the socket messages if the user is on or off and changing from app to server.