Update: index.html, update: index.js
Update index.html emitting a message from the frontend with the name 'mensaje' and a object that contain a name, also reading a message from the server(backend).
Update index.js reading the message from the frontend and show it in the terminal, also emitting a message.
Those two connections are made to test the comunication between each other, as the backend to the frontend and vice versa.